Rest_api_tests/requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
attr==0.3.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
distlib==0.3.6
execnet==1.9.0
Faker==14.2.0
filelock==3.8.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
virtualenv==20.16.5